@@ -129,6 +129,9 @@ class IndexAxis(AxisInterface):
         assert stop <= self.count
         return IndexAxis(self.name, stop - start), sl
 
+    def __eq__(self, other):
+        return self.count == other.count
+
     def intersection(self, friend, return_slices=False):
         count_out = min(self.count, friend.count)
         ax = IndexAxis(self.name, count_out)
@@ -190,6 +193,11 @@ class OffsetAxis(AxisInterface):
         assert stop <= self.offset + self.count
         return (OffsetAxis(self.name, stop - start, start, self.origin_tag),
                 slice(start - self.offset, stop - self.offset, stride))
+
+    def __eq__(self, other):
+        return (self.count == other.count and
+                self.offset == other.offset and
+                self.origin_tag == other.origin_tag)
 
     def intersection(self, friend, return_slices=False):
         offset = max(self.offset, friend.offset)
@@ -267,6 +275,10 @@ class LabelAxis(AxisInterface):
         _, i0, i1 = get_coindices(selector, self.vals)
         assert len(i0) == len(selector)  # not a strict subset!
         return LabelAxis(self.name, selector), i1
+
+    def __eq__(self, other):
+        return (self.count == other.count and
+                np.all(self.vals == other.vals))
 
     def intersection(self, friend, return_slices=False):
         _vals, i0, i1 = get_coindices(self.vals, friend.vals)
@@ -349,29 +361,63 @@ class AxisManager:
             self._fields[new_name] = self._fields.pop(name)
             self._assignments[new_name] = self._assignments.pop(name)
         return self
-    
+
     def add_axis(self, a):
         assert isinstance( a, AxisInterface)
         self._axes[a.name] = a.copy()
 
     def __contains__(self, name):
-        return name in self._fields or name in self._axes
+        attrs = name.split(".")
+        tmp_item = self
+        while attrs:
+            attr_name = attrs.pop(0)
+            if attr_name in tmp_item._fields:
+                tmp_item = tmp_item._fields[attr_name]
+            elif attr_name in tmp_item._axes:
+                tmp_item = tmp_item._axes[attr_name]
+            else:
+                return False
+        return True
 
     def __getitem__(self, name):
-        if name in self._fields:
-            return self._fields[name]
-        if name in self._axes:
-            return self._axes[name]
-        raise KeyError(name)
+
+        # We want to support options like:
+        # aman.focal_plane.xi . aman['focal_plane.xi']
+        # We will safely assume that a getitem will always have '.' as the separator
+        attrs = name.split(".")
+        tmp_item = self
+        while attrs:
+            attr_name = attrs.pop(0)
+            if attr_name in tmp_item._fields:
+                tmp_item = tmp_item._fields[attr_name]
+            elif attr_name in tmp_item._axes:
+                tmp_item = tmp_item._axes[attr_name]
+            else:
+                raise KeyError(attr_name)
+        return tmp_item
 
     def __setitem__(self, name, val):
-        if name in self._fields:
-            self._fields[name] = val
+
+        last_pos = name.rfind(".")
+        val_key = name
+        tmp_item = self
+        if last_pos > -1:
+            val_key = name[last_pos + 1:]
+            attrs = name[:last_pos]
+            tmp_item = self[attrs]
+
+        if isinstance(val, AxisManager) and isinstance(tmp_item, AxisManager):
+            raise ValueError("Cannot assign AxisManager to AxisManager. Please use wrap method.")
+
+        if val_key in tmp_item._fields:
+            tmp_item._fields[val_key] = val
         else:
-            raise KeyError(name)
+            raise KeyError(val_key)
 
     def __setattr__(self, name, value):
         # Assignment to members update those members
+        # We will assume that a path exists until the last member.
+        # If any member prior to that does not exist a keyerror is raised.
         if "_fields" in self.__dict__ and name in self._fields.keys():
             self._fields[name] = value
         else:
@@ -381,7 +427,11 @@ class AxisManager:
     def __getattr__(self, name):
         # Prevent members from override special class members.
         if name.startswith("__"): raise AttributeError(name)
-        return self[name]
+        try:
+            val = self[name]
+        except KeyError as ex:
+            raise AttributeError(name) from ex
+        return val
 
     def __dir__(self):
         return sorted(tuple(self.__dict__.keys()) + tuple(self.keys()))
@@ -514,12 +564,12 @@ class AxisManager:
                 output.wrap(k, new_data[k], axis_map)
             else:
                 if other_fields == "exact":
-                    ## if every item named k is a scalar 
+                    ## if every item named k is a scalar
                     err_msg = (f"The field '{k}' does not share axis '{axis}'; " 
                               f"{k} is not identical across all items " 
                               f"pass other_fields='drop' or 'first' or else " 
                               f"remove this field from the targets.")
-                                
+
                     if np.any([np.isscalar(i[k]) for i in items]):
                         if not np.all([np.isscalar(i[k]) for i in items]):
                             raise ValueError(err_msg)
@@ -527,14 +577,14 @@ class AxisManager:
                             raise ValueError(err_msg)
                         output.wrap(k, items[0][k], axis_map)
                         continue
-                        
+
                     elif not np.all([i[k].shape==items[0][k].shape for i in items]):
                         raise ValueError(err_msg)
                     elif not np.all([np.array_equal(i[k], items[0][k], equal_nan=True) for i in items]):
                         raise ValueError(err_msg)
-                        
+
                     output.wrap(k, items[0][k].copy(), axis_map)
-                    
+
                 elif other_fields == 'fail':
                     raise ValueError(
                         f"The field '{k}' does not share axis '{axis}'; "
@@ -553,7 +603,7 @@ class AxisManager:
     # Add and remove data while maintaining internal consistency.
 
     def wrap(self, name, data, axis_map=None,
-             overwrite=False):
+             overwrite=False, restrict_in_place=False):
         """Add data into the AxisManager.
 
         Arguments:
@@ -574,6 +624,11 @@ class AxisManager:
             
           overwrite (bool): If True then will write over existing data
             in field ``name`` if present.
+
+          restrict_in_place (bool): If True, then a wrapped
+            AxisManager may be modified and added, without a copy
+            first.  This can be much faster, if there's no need to
+            preserve the wrapped item.
 
         """
         if overwrite and (name in self._fields):
@@ -617,7 +672,7 @@ class AxisManager:
                 assign[index] = axis.name
         helper._fields[name] = data
         helper._assignments[name] = assign
-        return self.merge(helper)
+        return self.merge(helper, restrict_in_place=restrict_in_place)
 
     def wrap_new(self, name, shape=None, cls=None, **kwargs):
         """Create a new object and wrap it, with axes mapped.  The shape can
@@ -694,6 +749,8 @@ class AxisManager:
         # If simple list/tuple of Axes is passed in, convert to dict
         if not isinstance(axes, dict):
             axes = {ax.name: ax for ax in axes}
+        axes = {k: v for k, v in axes.items()
+                if k in dest._axes and dest._axes[k] != v}
         for name, ax in axes.items():
             if name not in dest._axes:
                 continue
@@ -705,7 +762,10 @@ class AxisManager:
             dest._axes[ax.name] = ax
         for k, v in self._fields.items():
             if isinstance(v, AxisManager):
-                dest._fields[k] = v.restrict_axes(axes, in_place=in_place)
+                if len(axes) == 0 and in_place:
+                    dest._fields[k] = v
+                else:
+                    dest._fields[k] = v.restrict_axes(axes, in_place=in_place)
             elif np.isscalar(v) or v is None:
                 dest._fields[k] = v
             else:
@@ -807,14 +867,19 @@ class AxisManager:
                     continue
                 if ax.name not in axes_out:
                     axes_out[ax.name] = ax.copy()
-                else:
+                elif axes_out[ax.name] != ax:
                     axes_out[ax.name] = axes_out[ax.name].intersection(
                         ax, False)
         return axes_out
 
-    def merge(self, *amans):
+    def merge(self, *amans, restrict_in_place=False):
         """Merge the data from other AxisMangers into this one.  Axes with the
         same name will be intersected.
+
+        If restrict_in_place=True, then the amans may be modified as
+        they are added to the output objcet.  When that arg is False,
+        the incoming amans are all copied, even if no modifications
+        are needed.
 
         """
         # Before messing with anything, check for key interference.
@@ -833,7 +898,7 @@ class AxisManager:
         self.restrict_axes(axes_out)
         # Import the other ones.
         for aman in amans:
-            aman = aman.restrict_axes(axes_out, in_place=False)
+            aman = aman.restrict_axes(axes_out, in_place=restrict_in_place)
             for k, v in aman._axes.items():
                 if k not in self._axes:
                     self._axes[k] = v
