@@ -13,7 +13,12 @@ import os
 import logging
 import sys
 import shutil
+import yaml
+import datetime as dt
+from zipfile import ZipFile
+import sotodlib
 from sotodlib.site_pipeline.util import init_logger
+from .datapkg_utils import walk_files
 
 
 log = logging.getLogger('bookbinder')
@@ -26,6 +31,10 @@ class TimingSystemOff(Exception):
 
 class NoScanFrames(Exception):
     """Exception raised when we try and bind a book but the SMuRF file contains not Scan frames (so no detector data)"""
+    pass
+
+class NoHKFiles(Exception):
+    """Exception raised when we cannot find any HK data around the book time"""
     pass
 
 class NoMountData(Exception):
@@ -68,13 +77,11 @@ def setup_logger(logfile=None):
 
     return log
 
-
 def get_frame_iter(files):
     """
     Returns a continuous iterator over frames for a list of files.
     """
     return itertools.chain(*[core.G3File(f) for f in files])
-
 
 def close_writer(writer):
     """
@@ -84,7 +91,6 @@ def close_writer(writer):
     if writer is None:
         return
     writer(core.G3Frame(core.G3FrameType.EndProcessing))
-
 
 def next_scan(it):
     """
@@ -97,7 +103,6 @@ def next_scan(it):
             return frame, interm_frames
         interm_frames.append(frame)
     return None, interm_frames
-
 
 class HkDataField:
     """
@@ -270,6 +275,13 @@ class AncilProcessor:
             self.log = logging.getLogger('bookbinder')
         else:
             self.log = log
+
+        if len(self.files) == 0:
+            if self.require_acu or self.require_hwp:
+                raise NoHKFiles("No HK files specified for book")
+            self.log.warning("No HK files found for book")
+            for fld in ['az', 'el', 'boresight', 'corotator_enc','az_mode', 'hwp_freq']:
+                setattr(self.hkdata, fld, None)
 
         if self.require_acu and self.hkdata.az is None:
             self.log.warning("No ACU data specified in hk_fields!")
@@ -480,7 +492,6 @@ class AncilProcessor:
             if k not in frame:
                 frame[k] = np.nan
 
-
 class SmurfStreamProcessor:
     def __init__(self, obs_id, files, book_id, readout_ids,
                  log=None, allow_bad_timing=False):
@@ -624,6 +635,10 @@ class SmurfStreamProcessor:
         # Sample idx within each in-frame for every input sample
         in_offset_idxs = np.arange(len(self.times)) - offsets[self.frame_idxs]
 
+        # Mask to use to separate out non-resonator signal from resonator signal
+        tracked_chans = np.array(['NONE' not in r for r in self.readout_ids], dtype=bool)
+        readout_id_arr = np.array(self.readout_ids)
+
         # Handle file writers
         writer = None
         cur_file_idx = None
@@ -740,7 +755,8 @@ class SmurfStreamProcessor:
             oframe['flag_smurfgaps'] = core.G3VectorBool(~filled)
 
             ts = core.G3VectorTime(ts * core.G3Units.s)
-            oframe['signal'] = so3g.G3SuperTimestream(self.readout_ids, ts, data)
+            oframe['signal'] = so3g.G3SuperTimestream(readout_id_arr[tracked_chans], ts, data[tracked_chans, :])
+            oframe['untracked'] = so3g.G3SuperTimestream(readout_id_arr[~tracked_chans], ts, data[~tracked_chans, :])
             oframe['primary'] = so3g.G3SuperTimestream(self.primary_names, ts, primary)
             oframe['tes_biases'] = so3g.G3SuperTimestream(self.bias_names, ts, biases)
             oframe['stream_id'] = self.stream_id
@@ -757,10 +773,10 @@ class SmurfStreamProcessor:
         if pbar.n >= pbar.total:
             pbar.close()
 
-
 class BookBinder:
     """
-    Class for combining smurf and hk L2 data to create books.
+    Class for combining smurf and hk L2 data to create books containing detector
+    timestreams.
 
     Parameters
     ----------
@@ -820,11 +836,11 @@ class BookBinder:
         self.data_root = data_root
         self.hk_root = os.path.join(data_root, 'hk')
         self.meta_root = os.path.join(data_root, 'smurf')
-        self.hkfiles = get_hk_files(self.hk_root, 
-                                    book.start.timestamp(),
-                                    book.stop.timestamp())
+        
         self.obsdb = obsdb
         self.outdir = outdir
+
+        assert book.schema==0, "obs/oper books only have schema=0"
 
         self.max_samps_per_frame = max_samps_per_frame
         self.max_file_size = max_file_size
@@ -850,6 +866,24 @@ class BookBinder:
 
         logfile = os.path.join(outdir, 'Z_bookbinder_log.txt')
         self.log = setup_logger(logfile)
+
+        try:
+            self.hkfiles = get_hk_files(
+                self.hk_root, 
+                book.start.timestamp(),
+                book.stop.timestamp()
+            )
+        except NoHKFiles as e:
+            if require_hwp or require_acu:
+                self.log.error(
+                    "HK files are required if we require ACU or HWP data"
+                )
+                raise e
+            self.log.warning(
+                "Found no HK files during book time, binding anyway because "
+                "require_acu and require_hwp are False"
+            )
+            self.hkfiles = []            
 
         self.ancil = AncilProcessor(
             self.hkfiles, 
@@ -966,6 +1000,34 @@ class BookBinder:
                 meta_files['noise'] = relpath
 
         self.meta_files = meta_files
+
+    def write_M_files(self, telescope, tube_config):
+        # write M_book file
+        m_book_file = os.path.join(self.outdir, "M_book.yaml")
+        book_meta = {}
+        book_meta["book"] = {
+            "type": self.book.type,
+            "schema_version": self.book.schema,
+            "book_id": self.book.bid,
+            "finalized_at": dt.datetime.utcnow().isoformat(),
+        }
+        book_meta["bookbinder"] = {
+            "codebase": sotodlib.__file__,
+            "version": sotodlib.__version__,
+            # leaving this in but KH doesn't know what it's supposed to be for
+            "context": "unknown", 
+        }
+        with open(m_book_file, "w") as f:
+            yaml.dump(book_meta, f)
+        
+        mfile = os.path.join(self.outdir, "M_index.yaml")
+        with open(mfile, "w") as f:
+            yaml.dump(
+                self.get_metadata(
+                    telescope=telescope,
+                    tube_config=tube_config,
+                ), f
+            )
 
     def get_metadata(self, telescope=None, tube_config={}):
         """
@@ -1091,6 +1153,108 @@ class BookBinder:
         self.log.info("Finished binding data. Exiting.")
         return True
 
+class TimeCodeBinder:
+    """Class for building the timecode based books, smurf, stray, and hk books. 
+    These books are built primarily just by copying specified files from level 
+    2 locations to new locations at level 2.
+    """
+
+    def __init__(
+        self, book, timecode, indir, outdir, file_list=None, 
+        ignore_pattern=None,
+    ):
+        self.book = book
+        self.timecode = timecode
+        self.indir = indir
+        self.outdir = outdir
+        self.file_list = file_list
+        if ignore_pattern is not None:
+            self.ignore_pattern = ignore_pattern
+        else:
+            self.ignore_pattern = []
+        
+        if book.type == 'smurf' and book.schema > 0:
+            self.compress_output = True
+        else:
+            self.compress_output = False    
+
+    def get_metadata(self, telescope=None, tube_config={}):
+        return {
+            "book_id": self.book.bid,
+            # dummy start and stop times
+            "start_time": float(self.timecode) * 1e5,
+            "stop_time": (float(self.timecode) + 1) * 1e5,
+            "telescope": telescope,
+            "type": self.book.type,
+        }
+    
+    def write_M_files(self, telescope, tube_config):
+        # write M_book file
+        
+        book_meta = {}
+        book_meta["book"] = {
+            "type": self.book.type,
+            "schema_version": self.book.schema,
+            "book_id": self.book.bid,
+            "finalized_at": dt.datetime.utcnow().isoformat(),
+        }
+        book_meta["bookbinder"] = {
+            "codebase": sotodlib.__file__,
+            "version": sotodlib.__version__,
+            # leaving this in but KH doesn't know what it's supposed to be for
+            "context": "unknown", 
+        }
+        if self.compress_output:
+            with ZipFile(self.outdir, mode='a') as zf:
+                zf.writestr("M_book.yaml", yaml.dump(book_meta))
+        else:
+            m_book_file = os.path.join(self.outdir, "M_book.yaml")
+            with open(m_book_file, "w") as f:
+                yaml.dump(book_meta, f)
+        
+        index = self.get_metadata(
+            telescope=telescope,
+            tube_config=tube_config,
+        )
+        if self.compress_output:
+            with ZipFile(self.outdir, mode='a') as zf:
+                zf.writestr("M_index.yaml", yaml.dump(index))
+        else:
+            mfile = os.path.join(self.outdir, "M_index.yaml")
+            with open(mfile, "w") as f:
+                yaml.dump(index, f)
+
+    def bind(self, pbar=False):
+        if self.compress_output:
+            if self.file_list is None:
+                self.file_list = walk_files(self.indir, include_suprsync=True)
+                ignore = shutil.ignore_patterns(*self.ignore_pattern)
+                to_ignore = ignore("", self.file_list)
+                self.file_list = sorted(
+                    [f for f in self.file_list if f not in to_ignore]
+                )
+            with ZipFile(self.outdir, mode='x') as zf:
+                for f in self.file_list:
+                    relpath = os.path.relpath(f, self.indir)
+                    zf.write(f, arcname=relpath)
+        elif self.file_list is None:
+            shutil.copytree(
+                self.indir,
+                self.outdir,
+                ignore=shutil.ignore_patterns(
+                    *self.ignore_pattern,
+                ),
+            )
+        else:
+            if not os.path.exists(self.outdir):
+                os.makedirs(self.outdir)
+            for f in self.file_list:
+                relpath = os.path.relpath(f, self.indir)
+                path = os.path.join(self.outdir, relpath)
+                base, _ = os.path.split(path)
+                if not os.path.exists(base):
+                    os.makedirs(base)
+                shutil.copy(f, os.path.join(self.outdir, relpath))
 
 def fill_time_gaps(ts):
     """
@@ -1132,7 +1296,6 @@ def fill_time_gaps(ts):
 
     return new_ts, ~m
 
-
 _primary_idx_map = {}
 def get_frame_times(frame, allow_bad_timing=False):
     """
@@ -1172,7 +1335,6 @@ def get_frame_times(frame, allow_bad_timing=False):
         ## don't change this error message. used in Imprinter CLI
         raise TimingSystemOff("Timing counters not incrementing")
 
-
 def split_ts_bits(c):
     """
     Split up 64 bit to 2x32 bit
@@ -1183,7 +1345,6 @@ def split_ts_bits(c):
     b = c & MAXINT
     return a, b
 
-
 def counters_to_timestamps(c0, c2):
     s, ns = split_ts_bits(c2)
 
@@ -1192,7 +1353,6 @@ def counters_to_timestamps(c0, c2):
     c2 = s + ns*1e-9 + 5*(4*365 + 1)*24*60*60
     ts = np.round(c2 - (c0 / 480000) ) + c0 / 480000
     return ts
-
 
 def find_ref_idxs(refs, vs):
     """
@@ -1251,8 +1411,10 @@ def get_hk_files(hkdir, start, stop, tbuff=10*60):
     m = (start-tbuff <= file_times) & (file_times < stop+tbuff)
     if not np.any(m):
         check = np.where( file_times <= start )
-        if len(check) < 1:
-            raise ValueError("Cannot find HK files we need")
+        if len(check) < 1 or len(check[0]) < 1:
+            raise NoHKFiles(
+                f"Cannot find HK files between {start} and {stop}"
+            )
         fidxs = [check[0][-1]]
         m[fidxs] = 1
     else:
@@ -1370,7 +1532,6 @@ def find_frame_splits(ancil, t0=None, t1=None):
     )
     idxs = locate_scan_events(az.times[msk], az.data[msk], filter_window=100)
     return az.times[msk][idxs]
-
 
 def get_smurf_files(obs, meta_path, all_files=False):
     """
