@@ -20,6 +20,9 @@ The config file could be of the form:
     lat_tube_list_file: path to yaml dict matching tubes and bands
     tolerate_stray_files: True
     skip_bad_books: True
+    known_bad_books: 
+    - oper_1736874485_satp3_0000100
+    - obs_9999999999_satp0_1111111
     extra_extra_files:
     - Z_bookbinder_log.txt
     extra_files:
@@ -30,7 +33,7 @@ The config file could be of the form:
 
 from sotodlib.core.metadata import ObsDb
 from sotodlib.core import Context 
-from sotodlib.site_pipeline.check_book import main as checkbook
+from sotodlib.site_pipeline import check_book
 from sotodlib.io import load_book
 import os
 import glob
@@ -42,6 +45,7 @@ import argparse
 import logging
 from sotodlib.site_pipeline import util
 from typing import Optional
+from itertools import product
 
 logger = util.init_logger('update_obsdb', 'update-obsdb: ')
 
@@ -86,7 +90,8 @@ def main(config: str,
         recency: float = None, 
         booktype: Optional[str] = "both",
         verbosity: Optional[int] = 2,
-        overwrite: Optional[bool] = False):
+        overwrite: Optional[bool] = False,
+        fastwalk: Optional[bool] = False):
 
     """
     Create or update an obsdb for observation or operations data.
@@ -104,6 +109,10 @@ def main(config: str,
         Output verbosity. 0:Error, 1:Warning, 2:Info(default), 3:Debug
     overwrite : bool
         if False, do not re-check existing entries
+    fastwalk : bool
+        if True, assume the directories have a structure /base_dir/obs|oper/\d{5}/...
+        Then replace base_dir with only the directories where \d{5} is greater or 
+        equal to recency.
     """
     if verbosity == 0:
         logger.setLevel(logging.ERROR)
@@ -144,6 +153,8 @@ def main(config: str,
         bookcartobsdb.add_obs_columns(col_list)
     if "skip_bad_books" not in config_dict:
         config_dict["skip_bad_books"] = False
+    if "known_bad_books" not in config_dict:
+        config_dict["known_bad_books"] = []
         
     #How far back we should look
     tnow = time.time()
@@ -156,6 +167,14 @@ def main(config: str,
     #Check if there are one or multiple base_dir specified
     if isinstance(base_dir,str):
         base_dir = [base_dir]
+    if fastwalk:
+        abv_tback = int(f"{int(tback):05}"[:5]) #Make sure we have at least five chars
+        abv_tnow = int(f"{int(tnow):05}"[:5])
+        abv_codes = np.arange(abv_tback, abv_tnow+1)
+        #Build the combinations base_dir/booktype/\d{5}
+        base_dir = [f"{os.path.join(x[0], x[1], str(x[2]))}" for x in product(base_dir, accept_type, abv_codes)]
+        logger.info(f"Looking in the following directories only: {str(base_dir)}")
+
     for bd in base_dir:
         #Find folders that are book-like and recent
         for dirpath, _, _ in os.walk(bd):
@@ -163,9 +182,12 @@ def main(config: str,
                 _, book_id = os.path.split(dirpath)
                 if book_id in existing and not overwrite:
                     continue
-                found_timestamp = re.search(r"\d{10}", book_id)#Find the rough timestamp
-                if found_timestamp and int(found_timestamp.group())>tback:
-                    #Looks like a book folder and young enough
+                if book_id in config_dict["known_bad_books"]:
+                    logger.debug(f"{book_id} known to be bad, skipping it")
+                    continue
+                edit_time = os.path.getmtime(dirpath)
+                if edit_time > tback:
+                    #Looks like a book folder and edited recently enough
                     bookcart.append(dirpath)
     
     logger.info(f"Found {len(bookcart)} new books in {time.time()-tnow} s")
@@ -173,10 +195,16 @@ def main(config: str,
     for bookpath in sorted(bookcart):
         if check_meta_type(bookpath) in accept_type:
             t1 = time.time()
+            logger.info(f"Examining book at {bookpath}")
             try:
                 #obsfiledb creation
-                checkbook(bookpath, config, add=True, overwrite=True)
-                logger.info(f"Ran check_book for {bookpath} in {time.time()-t1} s")
+                ok, obsfiledb_info = check_book.scan_book_dir(
+                    bookpath, logger, config_dict, prep_obsfiledb=True)
+                if not ok:
+                    raise RuntimeError("check_book found fatal errors, not adding.")
+                check_book.add_to_obsfiledb(
+                    obsfiledb_info, logger, config_dict, overwrite=True)
+                logger.info(f"Ran check_book in {time.time()-t1} s")
             except Exception as e:
                 if config_dict["skip_bad_books"]:
                     logger.warning(f"failed to add {bookpath}")
@@ -295,7 +323,7 @@ def main(config: str,
             tags = [t.strip() for t in tags if t.strip() != '']
 
             bookcartobsdb.update_obs(obs_id, very_clean, tags=tags)
-            logger.info(f"Added {obs_id} in {time.time()-t1} s")
+            logger.info(f"Finished {obs_id} in {time.time()-t1} s")
         else:
             bookcart.remove(bookpath)
 
@@ -305,7 +333,7 @@ def get_parser(parser=None):
         parser = argparse.ArgumentParser()
     parser.add_argument("--config", help="ObsDb, ObsfileDb configuration file", 
         type=str, required=True)
-    parser.add_argument('--recency', default=None, type=float,
+    parser.add_argument("--recency", default=None, type=float,
         help="Days to subtract from now to set as minimum ctime. If None, no minimum")
     parser.add_argument("--verbosity", default=2, type=int,
         help="Increase output verbosity. 0:Error, 1:Warning, 2:Info(default), 3:Debug")
@@ -313,6 +341,8 @@ def get_parser(parser=None):
         help="Select book type to look for: obs, oper, both(default)")
     parser.add_argument("--overwrite", action="store_true",
         help="If true, writes over existing entries")
+    parser.add_argument("--fastwalk", action="store_true",
+        help="Assume known directory tree shape and speed up walkthrough")
     return parser
 
 
