@@ -2,21 +2,20 @@
 # Full license can be found in the top level "LICENSE" file.
 """Helper functions for LoadContext operator."""
 
-import os
 import numpy as np
 
-from astropy import units as u
-from astropy.table import Column, QTable
+from astropy.table import Column
 
 import so3g
 
 import toast
 from toast.mpi import MPI
 from toast.timing import function_timer, Timer
-from toast.utils import Environment, Logger
-from toast.dist import distribute_discrete, distribute_uniform
+from toast.utils import Logger
+from toast.dist import distribute_uniform
+from toast.observation import default_values as defaults
 
-from ...core import Context, AxisManager, FlagManager
+from ...core import Context, AxisManager
 from ...core.axisman import AxisInterface
 from ...preprocess import Pipeline as PreProcPipe
 from ...hwp.hwp_angle_model import apply_hwp_angle_model
@@ -59,7 +58,7 @@ def read_and_preprocess_wafers(
         context_file (str):  The context file to open or None.
 
     Returns:
-        (dict):  The AxisManager data for each wafer on this process (or None).
+        (dict):  The AxisManager data for each wafer on this process.
 
     """
     log = Logger.get()
@@ -177,12 +176,40 @@ def parse_metadata(axman, obs_meta, fp_cols, path_sep, det_axis, obs_base, fp_ba
         del obs_meta[obs_base]
 
 
+def _local_process_dets(obs, rank):
+    """Get the list of local detectors for a specific rank."""
+    # Full detector list
+    det_names = obs.all_detectors
+    # The range of full detector indices on this process.
+    full_indices = obs.dist.det_indices[rank]
+    # The same info, as a slice.
+    full_slc = slice(full_indices.offset, full_indices.offset + full_indices.n_elem, 1)
+    # The names of the detectors on this process.
+    return det_names[full_slc]
+
+
 def distribute_detector_props(obs, wafer_key):
     """Compute the communication patterns for detector data.
 
     This computes the range of local detector indices for each wafer on all
     processes, and also the corresponding range of indices in the wafer data
-    on the reading process.
+    on the reading process.  For example, imagine we have 2 wafers and 4
+    processes:
+
+    +====================+    +====================+
+    |       Wafer 0      |    |     Process 0      |
+    |                    |    |--------------------|
+    |                    |    |     Process 1      |
+    +====================+    |                    |
+    |       Wafer 1      |    |--------------------|
+    |                    |    |     Process 2      |
+    |                    |    |--------------------|
+    |                    |    |     Process 3      |
+    +====================+    +--------------------+
+
+    In this case, Process 0 will read Wafer 0 and Process 1 will read Wafer 1.
+    The data from Wafer 0 needs to be sent to Processes 0 and 1, and the data
+    from Wafer 1 needs to be sent to Processes 1, 2, and 3.
 
     Args:
         obs (Observation):  The observation.
@@ -197,12 +224,25 @@ def distribute_detector_props(obs, wafer_key):
     gsize = obs.comm.group_size
 
     det_table = obs.telescope.focalplane.detector_data
-    det_names = obs.all_detectors
+
+    # Wafer names for all detectors
     det_wafers = list(det_table[wafer_key])
 
+    # For each process, this has a key for each wafer on that process.
+    # The value for each wafer key is the destination local detector indices for
+    # that wafer.
     proc_wafer_dets = dict()
+
+    # For each wafer, this has a key for each destination process.  The value
+    # for each process key is the range of detector indices in the wafer data
+    # that should be sent to that process.
     wafer_proc_dets = dict()
+
+    # For each wafer, the process that should do the reading.  This is the first
+    # process rank that has any data from that wafer.
     wafer_readers = dict()
+
+    # For each wafer, the list of local detector names within that wafer.
     wafer_dets = dict()
 
     wafer_off = 0
@@ -212,13 +252,17 @@ def distribute_detector_props(obs, wafer_key):
     wafer_dets[cur_wafer] = list()
     for proc in range(gsize):
         proc_wafer_dets[proc] = dict()
+        # The range of full detector indices on this process.
         full_indices = obs.dist.det_indices[proc]
+        # The same info, as a slice.
         full_slc = slice(
             full_indices.offset, full_indices.offset + full_indices.n_elem, 1
         )
-        local_det_names = det_names[full_slc]
+        # The names of the detectors on this process.
+        local_det_names = _local_process_dets(obs, proc)
+        # The wafer names of the detectors on this process.
         local_det_wafers = det_wafers[full_slc]
-        local_wafers = list(sorted(set(local_det_wafers)))
+
         if local_det_wafers[0] != cur_wafer:
             # This process starts on a new wafer
             cur_wafer = local_det_wafers[0]
@@ -265,6 +309,8 @@ def distribute_detector_data(
     obs,
     field,
     axwafers,
+    axis_dets,
+    axis_samples,
     axfield,
     axdtype,
     wafer_readers,
@@ -274,6 +320,40 @@ def distribute_detector_data(
     flag_invert=False,
     flag_mask=None,
 ):
+    """Communicate detector data from the reading processes to the destinations.
+
+    Data for each wafer is loaded on exactly one process.  This function uses the
+    pre-computed data communication pattern built by `distribute_detector_props()`.
+    However, the loaded wafer data AxisManager may have been truncated with a
+    symmetric sample buffer on either side and some detectors may have been removed.
+    This function must account for this when copying data from the AxisManager into
+    The full-size observation.
+
+    The observation data is modified in-place.
+
+    Args:
+        obs (Observation):  The observation.
+        field (str):  The detdata field in the Observation.
+        axwafers (dict):  The dictionary of wafer data loaded on this process.
+        axis_dets (str):  The name of the detector LabelAxis in the AxisManagers.
+        axis_samples (str):  The name of the sample OffsetAxis in the AxisManagers.
+        axfield (str):  The name of the data field in the AxisManagers.
+        axdtype (np.dtype):  The dtype of the source buffer for sending the data
+            from the AxisManager.
+        wafer_readers (dict):  The reading process for each wafer.
+        wafer_proc_dets (dict):  For each wafer, the dictionary of destination
+            processes and the detector indices to send.
+        proc_wafer_dets (dict):  For each process, the dictionary of wafer name to
+            local detector indices to receive.
+        is_flag (bool):  If True, this field is a flag.
+        flag_invert (bool):  If True, invert the meaning of the flag values.
+        flag_mask (np.uint8):  The flag mask (or None).
+
+    Returns:
+        None
+
+    """
+    log = Logger.get()
     gcomm = obs.comm.comm_group
     rank = obs.comm.group_rank
     gsize = obs.comm.group_size
@@ -304,47 +384,112 @@ def distribute_detector_data(
 
     # The tag value stride, to ensure unique tags for every sender / receiver / wafer
     # combination.
-    tag_stride = 1000
+    tag_stride = 2000
 
     wf_index = {y: x for x, y in enumerate(wafer_readers.keys())}
 
     for wafer, reader in wafer_readers.items():
         if reader == rank:
             send_data[wafer] = dict()
+            # The wafer axis manager may have a different set of detectors than
+            # the full data, if some detectors have been cut.  It may also have
+            # a symmetric buffer of samples cut on either side of the observation.
+            # here we construct the mapping from full data into this restricted
+            # axis manager.
+            restricted_dets = axwafers[wafer][axis_dets].vals
+            restricted_indices = {y: x for x, y in enumerate(restricted_dets)}
+
             for receiver, send_dets in wafer_proc_dets[wafer].items():
+                # "send_dets" is the un-restricted range of wafer detectors.
+                # We will send a full-size (unrestricted) buffer to ease
+                # bookkeeping, and just copy un-cut detector data into that
+                # buffer.
+                n_send_det = send_dets[1] - send_dets[0]
+
+                # The names of the detectors on the receiving process.
+                local_det_names = _local_process_dets(obs, receiver)
+
+                # The receiving detector names and their relative indices
                 recv_dets = proc_wafer_dets[receiver][wafer]
+                n_recv_det = recv_dets[1] - recv_dets[0]
+                recv_det_names = local_det_names[recv_dets[0] : recv_dets[1]]
+                recv_det_indices = {y: x for x, y in enumerate(recv_det_names)}
+
+                # Build the mapping of restricted indices to send buffer indices
+                restrict_to_send = {
+                    restricted_indices[x]: y
+                    for x, y in recv_det_indices.items()
+                    if x in restricted_indices
+                }
+
+                # The per-detector flags, so that detectors are cut properly on the
+                # receiver.
+                det_flags = np.ones(
+                    n_send_det,
+                    dtype=np.uint8,
+                )
+                for idet_ax, idet_send in restrict_to_send.items():
+                    det_flags[idet_send] = 0
+
+                # Does the axis manager have a truncated number of samples?
+                restricted_samps = axwafers[wafer][axis_samples].count
+                if restricted_samps != obs.n_local_samples:
+                    ax_shift = axwafers[wafer][axis_samples].offset
+                else:
+                    ax_shift = 0
+
+                # flat-packed buffer size of send buffer
+                flat_size = n_send_det * obs.n_local_samples
+
                 # Is this some detector flag data using ranges instead of samples?
                 # If so, we construct a temporary buffer and build sample flags
                 # from the ranges.
-                n_send_det = send_dets[1] - send_dets[0]
-                flat_size = n_send_det * obs.n_local_samples
                 if isinstance(axwafers[wafer][axfield], so3g.proj.RangesMatrix):
-                    # Yes, flagged ranges
-                    sdata = np.empty(
+                    # Yes, flagged ranges.  We may have restricted sample ranges
+                    # for our flags, and so we initialize the full buffer to
+                    # the invalid mask.
+                    sdata = defaults.det_mask_invalid * np.ones(
                         flat_size,
                         dtype=np.uint8,
                     )
                     if flag_invert:
-                        sdata[:] = flag_mask
-                        for idet in range(n_send_det):
-                            off = idet * obs.n_local_samples
-                            for rg in axwafers[wafer][axfield][idet].ranges():
+                        for idet_ax, idet_send in restrict_to_send.items():
+                            # Set this detector's values to the mask, and then
+                            # we will "unflag" the specified ranges.
+                            off = idet_send * obs.n_local_samples + ax_shift
+                            sdata[off : off + restricted_samps] = flag_mask
+                            for rg in axwafers[wafer][axfield][idet_ax].ranges():
                                 sdata[off + rg[0] : off + rg[1]] = 0
                     else:
-                        sdata[:] = 0
-                        for idet in range(n_send_det):
-                            off = idet * obs.n_local_samples
-                            for rg in axwafers[wafer][axfield][idet].ranges():
+                        for idet_ax, idet_send in restrict_to_send.items():
+                            # Set this detector's values to good, and then
+                            # we will flag the specified ranges.
+                            off = idet_send * obs.n_local_samples + ax_shift
+                            sdata[off : off + restricted_samps] = 0
+                            for rg in axwafers[wafer][axfield][idet_ax].ranges():
                                 sdata[off + rg[0] : off + rg[1]] = flag_mask
                 else:
                     # Either normal sample flags or signal data
-                    sdata = np.empty(
-                        flat_size,
-                        dtype=axdtype,
-                    )
-                    sdata[:] = np.ravel(
-                        axwafers[wafer][axfield][send_dets[0] : send_dets[1], :]
-                    )
+                    if is_flag:
+                        # We may have restricted sample ranges for our flags, and so
+                        # we initialize the full buffer to the invalid mask.
+                        sdata = defaults.det_mask_invalid * np.ones(
+                            flat_size,
+                            dtype=axdtype,
+                        )
+                    else:
+                        # Signal data is initialized to zero.  Restricted / missing
+                        # samples will be indicated by flags.
+                        sdata = np.zeros(
+                            flat_size,
+                            dtype=axdtype,
+                        )
+                    for idet_ax, idet_send in restrict_to_send.items():
+                        off = idet_send * obs.n_local_samples + ax_shift
+                        sdata[off : off + restricted_samps] = axwafers[wafer][axfield][
+                            idet_ax, :
+                        ]
+
                 if receiver == rank:
                     # We just need to process the data locally
                     if is_flag:
@@ -358,38 +503,58 @@ def distribute_detector_data(
                         obs.detdata[field][recv_dets[0] : recv_dets[1], :] = (
                             sdata.reshape((n_send_det, -1))
                         )
+                        # Update per-detector flags
+                    dflags = {
+                        obs.local_detectors[recv_dets[0] + x]: defaults.det_mask_invalid
+                        for x in range(n_recv_det)
+                        if det_flags[x] != 0
+                    }
+                    obs.update_local_detector_flags(dflags)
                 else:
                     # Send asynchronously.
-                    tag = (rank * gsize + receiver) * tag_stride + wf_index[wafer]
+                    tag = (rank * gsize + receiver) * tag_stride + 2 * wf_index[wafer]
+                    flag_tag = tag + 1
                     req = gcomm.isend(sdata, dest=receiver, tag=tag)
+                    req_flags = gcomm.isend(det_flags, dest=receiver, tag=flag_tag)
                     # Save a handle to this buffer while send operation is in progress
                     send_data[wafer][axfield] = sdata
+                    send_data[wafer][f"{axfield}_flags"] = det_flags
                     send_req.append(req)
+                    send_req.append(req_flags)
 
     my_wafer_dets = proc_wafer_dets[rank]
     for wafer, recv_dets in my_wafer_dets.items():
+        n_recv_det = recv_dets[1] - recv_dets[0]
         sender = wafer_readers[wafer]
         if sender == rank:
             # This data was already copied locally above
             continue
         else:
             # Receive from sender.  We always allocate a contiguous temporary buffer.
-            tag = (sender * gsize + rank) * tag_stride + wf_index[wafer]
-            n_recv_det = recv_dets[1] - recv_dets[0]
+            tag = (sender * gsize + rank) * tag_stride + 2 * wf_index[wafer]
+            flag_tag = tag + 1
             recv_data = gcomm.recv(source=sender, tag=tag)
+            det_flags = gcomm.recv(source=sender, tag=flag_tag)
+            det_slc = slice(recv_dets[0], recv_dets[1], 1)
             if is_flag:
                 _process_flag(
                     flag_mask,
                     recv_data.reshape((n_recv_det, -1)),
-                    obs.detdata[field][recv_dets[0] : recv_dets[1], :],
+                    obs.detdata[field][det_slc, :],
                     flag_invert,
                 )
             else:
                 # Just assign
-                obs.detdata[field][recv_dets[0] : recv_dets[1], :] = recv_data.reshape(
-                    (n_recv_det, -1)
-                )
+                obs.detdata[field][det_slc, :] = recv_data.reshape((n_recv_det, -1))
+            # Update per-detector flags
+            dflags = {
+                obs.local_detectors[recv_dets[0] + x]: defaults.det_mask_invalid
+                for x in range(n_recv_det)
+                if det_flags[x] != 0
+            }
+            obs.update_local_detector_flags(dflags)
             del recv_data
+            del det_flags
 
     # Wait for communication
     for req in send_req:
