@@ -2,33 +2,35 @@ import os
 import unittest
 import json
 import shutil
-import pytest
 
-from sotodlib.core.resources import get_local_file
+from sotodlib.core import resources
+
+ODD_NAME = 'planet47-ephemeris.txt'
 
 
 class TestCoreResources(unittest.TestCase):
 
     def test_get_local_file(self):
 
-        # t = get_local_file("de421.bsp", cache=True)
-        # expected_path = os.path.join(
-        #     os.path.expanduser("~"), ".sotodlib/filecache/de421.bsp"
-        # )
-        # self.assertEqual(expected_path, t)
-        # shutil.rmtree(os.path.join(os.path.expanduser("~"), ".sotodlib/"))
+        # This url should not be accessed, as long as download=False
+        resources.RESOURCE_DEFAULTS[ODD_NAME] = \
+            'ftp://simonsobs.org/sotodlib-ci.txt'
+
         os.environ["SOTODLIB_RESOURCES"] = json.dumps(
             {"someotherfile": "file://somepath/otherfile"}
         )
-        t = get_local_file("de421.bsp", cache=False)
-        expected_path = "/tmp/de421.bsp"
+
+        # Use a fake file for this, or else it might find a
+        # user-cached copy.
+        t = resources.get_local_file(ODD_NAME, cache=False, download=False)
+        expected_path = "/tmp/" + ODD_NAME
         self.assertEqual(expected_path, t)
 
         os.environ["SOTODLIB_RESOURCES"] = json.dumps(
-            {"de421.bsp": "file://somepath/de421.bsp"}
+            {ODD_NAME: "file://somepath/de421.bsp"}
         )
-        t = get_local_file("de421.bsp")
+        t = resources.get_local_file(ODD_NAME)
         self.assertEqual("somepath/de421.bsp", t)
 
-        with pytest.raises(RuntimeError):
-            get_local_file("doesnotexist.file")
+        with self.assertRaises(RuntimeError):
+            resources.get_local_file("doesnotexist.file")
